Redesigned hook for blocking screen recording detection.
