+++ app/src/main/java/com/keshav/capturesposed/CaptureSposed.kt
@@ -1,7 +1,6 @@
 package com.keshav.capturesposed
 
 import android.os.Build
-import com.keshav.capturesposed.hookers.ScreenRecordingCallbackControllerHooker
 import com.keshav.capturesposed.hookers.SystemUIHooker
 import com.keshav.capturesposed.hookers.WindowManagerServiceHooker
 import io.github.libxposed.api.XposedInterface
@@ -22,9 +21,6 @@ class CaptureSposed(base: XposedInterface, param: ModuleLoadedParam) : XposedMod
 
         try {
             WindowManagerServiceHooker.hook(param, module)
-
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.VANILLA_ICE_CREAM)
-                ScreenRecordingCallbackControllerHooker.hook(param, module)
         } catch (e: Exception) {
             log("[CaptureSposed] ERROR: $e")
         }
